gpu-services/coqui/requirements.txt
fastapi
uvicorn[standard]
python-multipart
coqui-tts
transformers>=4.57,<5.0
soundfile
numpy

